requirements.txt
certifi==2023.5.7
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
filelock==3.12.0
fonttools==4.39.4
imageio==2.28.1
kiwisolver==1.4.4
lazy_loader==0.2
matplotlib==3.7.1
networkx==3.1
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pipenv==2023.4.29
platformdirs==3.5.1
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
tifffile==2023.4.12
virtualenv==20.23.0
virtualenv-clone==0.5.7
